feat(61): added price validation to two decimal places
