+++ .gitignore
@@ -24,3 +24,4 @@ TODO.md
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+.idea/
